safari-600.7-branch/LayoutTests/js/dfg-mod-neg2tothe31-by-one-and-then-or-zero-with-interesting-reg-alloc-expected.txt

Tests that the (-2147483648/-1)|0 case in the DFG is handled correctly even when there is some interesting register allocation going on.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS foo(0, 0, -2147483647-1, -1) is 0
PASS foo(0, 0, -2147483647-1, -1) is 0
PASS foo(0, 0, -2147483647-1, -1) is 0
PASS foo(0, 0, -2147483647-1, -1) is 0
PASS foo(0, 0, -2147483647-1, -1) is 0
PASS foo(0, 0, -2147483647-1, -1) is 0
PASS foo(0, 0, -2147483647-1, -1) is 0
PASS foo(0, 0, -2147483647-1, -1) is 0
PASS foo(0, 0, -2147483647-1, -1) is 0
PASS foo(0, 0, -2147483647-1, -1) is 0
PASS foo(0, 0, -2147483647-1, -1) is 0
PASS foo(0, 0, -2147483647-1, -1) is 0
PASS foo(0, 0, -2147483647-1, -1) is 0
PASS foo(0, 0, -2147483647-1, -1) is 0
PASS foo(0, 0, -2147483647-1, -1) is 0
PASS foo(0, 0, -2147483647-1, -1) is 0
PASS foo(0, 0, -2147483647-1, -1) is 0
PASS foo(0, 0, -2147483647-1, -1) is 0
PASS foo(0, 0, -2147483647-1, -1) is 0
PASS foo(0, 0, -2147483647-1, -1) is 0
PASS foo(0, 0, -2147483647-1, -1) is 0
PASS foo(0, 0, -2147483647-1, -1) is 0
PASS foo(0, 0, -2147483647-1, -1) is 0
PASS foo(0, 0, -2147483647-1, -1) is 0
PASS foo(0, 0, -2147483647-1, -1) is 0
PASS foo(0, 0, -2147483647-1, -1) is 0
PASS foo(0, 0, -2147483647-1, -1) is 0
PASS foo(0, 0, -2147483647-1, -1) is 0
PASS foo(0, 0, -2147483647-1, -1) is 0
PASS foo(0, 0, -2147483647-1, -1) is 0
PASS foo(0, 0, -2147483647-1, -1) is 0
PASS foo(0, 0, -2147483647-1, -1) is 0
PASS foo(0, 0, -2147483647-1, -1) is 0
PASS foo(0, 0, -2147483647-1, -1) is 0
PASS foo(0, 0, -2147483647-1, -1) is 0
PASS foo(0, 0, -2147483647-1, -1) is 0
PASS foo(0, 0, -2147483647-1, -1) is 0
PASS foo(0, 0, -2147483647-1, -1) is 0
PASS foo(0, 0, -2147483647-1, -1) is 0
PASS foo(0, 0, -2147483647-1, -1) is 0
PASS foo(0, 0, -2147483647-1, -1) is 0
PASS foo(0, 0, -2147483647-1, -1) is 0
PASS foo(0, 0, -2147483647-1, -1) is 0
PASS foo(0, 0, -2147483647-1, -1) is 0
PASS foo(0, 0, -2147483647-1, -1) is 0
PASS foo(0, 0, -2147483647-1, -1) is 0
PASS foo(0, 0, -2147483647-1, -1) is 0
PASS foo(0, 0, -2147483647-1, -1) is 0
PASS foo(0, 0, -2147483647-1, -1) is 0
PASS foo(0, 0, -2147483647-1, -1) is 0
PASS foo(0, 0, -2147483647-1, -1) is 0
PASS foo(0, 0, -2147483647-1, -1) is 0
PASS foo(0, 0, -2147483647-1, -1) is 0
PASS foo(0, 0, -2147483647-1, -1) is 0
PASS foo(0, 0, -2147483647-1, -1) is 0
PASS foo(0, 0, -2147483647-1, -1) is 0
PASS foo(0, 0, -2147483647-1, -1) is 0
PASS foo(0, 0, -2147483647-1, -1) is 0
PASS foo(0, 0, -2147483647-1, -1) is 0
PASS foo(0, 0, -2147483647-1, -1) is 0
PASS foo(0, 0, -2147483647-1, -1) is 0
PASS foo(0, 0, -2147483647-1, -1) is 0
PASS foo(0, 0, -2147483647-1, -1) is 0
PASS foo(0, 0, -2147483647-1, -1) is 0
PASS foo(0, 0, -2147483647-1, -1) is 0
PASS foo(0, 0, -2147483647-1, -1) is 0
PASS foo(0, 0, -2147483647-1, -1) is 0
PASS foo(0, 0, -2147483647-1, -1) is 0
PASS foo(0, 0, -2147483647-1, -1) is 0
PASS foo(0, 0, -2147483647-1, -1) is 0
PASS foo(0, 0, -2147483647-1, -1) is 0
PASS foo(0, 0, -2147483647-1, -1) is 0
PASS foo(0, 0, -2147483647-1, -1) is 0
PASS foo(0, 0, -2147483647-1, -1) is 0
PASS foo(0, 0, -2147483647-1, -1) is 0
PASS foo(0, 0, -2147483647-1, -1) is 0
PASS foo(0, 0, -2147483647-1, -1) is 0
PASS foo(0, 0, -2147483647-1, -1) is 0
PASS foo(0, 0, -2147483647-1, -1) is 0
PASS foo(0, 0, -2147483647-1, -1) is 0
PASS foo(0, 0, -2147483647-1, -1) is 0
PASS foo(0, 0, -2147483647-1, -1) is 0
PASS foo(0, 0, -2147483647-1, -1) is 0
PASS foo(0, 0, -2147483647-1, -1) is 0
PASS foo(0, 0, -2147483647-1, -1) is 0
PASS foo(0, 0, -2147483647-1, -1) is 0
PASS foo(0, 0, -2147483647-1, -1) is 0
PASS foo(0, 0, -2147483647-1, -1) is 0
PASS foo(0, 0, -2147483647-1, -1) is 0
PASS foo(0, 0, -2147483647-1, -1) is 0
PASS foo(0, 0, -2147483647-1, -1) is 0
PASS foo(0, 0, -2147483647-1, -1) is 0
PASS foo(0, 0, -2147483647-1, -1) is 0
PASS foo(0, 0, -2147483647-1, -1) is 0
PASS foo(0, 0, -2147483647-1, -1) is 0
PASS foo(0, 0, -2147483647-1, -1) is 0
PASS foo(0, 0, -2147483647-1, -1) is 0
PASS foo(0, 0, -2147483647-1, -1) is 0
PASS foo(0, 0, -2147483647-1, -1) is 0
PASS foo(0, 0, -2147483647-1, -1) is 0
PASS successfullyParsed is true

TEST COMPLETE
